Show recheduled salary request in UI

+++ platform/src/main/java/ru/abondin/hreasy/platform/service/salary/dto/link/SalaryRequestLinkCreateBody.java
@@ -17,6 +17,6 @@ public class SalaryRequestLinkCreateBody {
     @NonNull
     private Integer destination;
     @NonNull
-    private Integer type;
+    private Short type;
     private String comment;
 }
